refactor: KidInfoBox 디렉토리 이동
