+++ FE/src/components/Bottomsheet/KidInfoBox/kidInfoBox.module.scss
